Move DirectInput.* and DirectVideo.* to a Windows folder.
