+++ source/Debugger/Debug.cpp
@@ -43,7 +43,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "../Memory.h"
 #include "../NTSC.h"
 #include "../SoundCore.h"	// SoundCore_SetFade()
-#include "../DirectVideo.h"
+#include "../Windows/DirectVideo.h"
 #include "../Video.h"
 
 //	#define DEBUG_COMMAND_HELP  1
