+++ source/Keyboard.cpp
@@ -34,7 +34,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Pravets.h"
 #include "Tape.h"
 #include "YamlHelper.h"
-#include "DirectVideo.h" // Needed by TK3000 //e, to refresh the frame at each |Mode| change
+#include "Windows/DirectVideo.h" // Needed by TK3000 //e, to refresh the frame at each |Mode| change
 #include "Log.h"
 
 static BYTE asciicode[2][10] = {
